requirements.txt
streamlit==1.47.0
pandas==2.2.3
plotly==6.2.0
google-genai
PyMuPDF==1.26.4
python-docx==1.2.0
Pillow==11.1.0
pytesseract==0.3.13
reportlab==4.4.3
PyPDF2==3.0.1
openai==1.107.1